파일명 변경 및 인터페이스,Impl 테스트 코드 구현
