+++ src/main/java/jpabook/jpashop_v2/repository/MemberRepositoryCustom.java
@@ -1,0 +1,11 @@
+package jpabook.jpashop_v2.repository;
+
+import jpabook.jpashop_v2.dto.MemberSearchCondition;
+import jpabook.jpashop_v2.dto.MemberTeamDto;
+
+import java.util.List;
+
+public interface MemberRepositoryCustom {
+
+    List<MemberTeamDto> search(MemberSearchCondition condition);
+}
