+++ src/main/java/jpabook/jpashop_v2/repository/MemberRepositoryImpl.java
@@ -1,0 +1,55 @@
+package jpabook.jpashop_v2.repository;
+
+import com.querydsl.core.types.dsl.BooleanExpression;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import jpabook.jpashop_v2.domain.QMember;
+import jpabook.jpashop_v2.dto.MemberSearchCondition;
+import jpabook.jpashop_v2.dto.MemberTeamDto;
+import jpabook.jpashop_v2.dto.QMemberTeamDto;
+import org.springframework.util.StringUtils;
+
+import javax.persistence.EntityManager;
+import java.util.List;
+
+import static jpabook.jpashop_v2.domain.QTeam.team;
+import static jpabook.jpashop_v2.domain.QMember.member;
+import static org.springframework.util.StringUtils.*;
+
+public class MemberRepositoryImpl implements MemberRepositoryCustom {
+
+    private final JPAQueryFactory queryFactory;
+    public MemberRepositoryImpl(EntityManager em) {
+        this.queryFactory = new JPAQueryFactory(em);
+    }
+    @Override
+    //회원명, 팀명, 나이(ageGoe, ageLoe)
+    public List<MemberTeamDto> search(MemberSearchCondition condition) {
+
+        return queryFactory
+                .select(new QMemberTeamDto(
+                        member.id.as("memberId"),
+                        member.name,
+                        member.age,
+                        team.id.as("teamId"),
+                        team.name.as("teamName")))
+                .from(member)
+                .leftJoin(member.team, team)
+                .where(usernameEq(condition.getUsername()),
+                        teamNameEq(condition.getTeamName()),
+                        ageGoe(condition.getAgeGoe()),
+                        ageLoe(condition.getAgeLoe()))
+                .fetch();
+    }
+    private BooleanExpression usernameEq(String username) {
+        return isEmpty(username) ? null : member.name.eq(username);
+    }
+    private BooleanExpression teamNameEq(String teamName) {
+        return isEmpty(teamName) ? null : team.name.eq(teamName);
+    }
+    private BooleanExpression ageGoe(Integer ageGoe) {
+        return ageGoe == null ? null : member.age.goe(ageGoe);
+    }
+    private BooleanExpression ageLoe(Integer ageLoe) {
+        return ageLoe == null ? null : member.age.loe(ageLoe);
+    }
+}
