+++ src/test/java/jpabook/jpashop_v2/repository/MemberRepositoryTest.java
@@ -25,56 +25,39 @@ public class MemberRepositoryTest
     EntityManager em;
 
     @Test
-    public void test001()
+    public void basicQuerydslTest()
     {
-
-    }
-
-    @Test
-    public void basicQuerydslTest() {
         Member member = new Member("memberdummy001");
         memberRepository.save(member);
         Member findMember = memberRepository.findById(member.getId()).get();
         assertThat(findMember).isEqualTo(member);
-        List<Member> result1 = memberRepository.findAll_Querydsl();
+        List<Member> result1 = memberRepository.findAll();
+
+        System.out.println("result 1 : "+result1.toString());
+        assertThat(result1).contains(member);
+    }
+
+    @Test
+    public void searchTest()
+    {
+        Member member = new Member("memberdummy001");
+        memberRepository.save(member);
+        Member findMember = memberRepository.findById(member.getId()).get();
+        assertThat(findMember).isEqualTo(member);
+        List<Member> result1 = memberRepository.findAll();
 
         System.out.println("result 1 : "+result1.toString());
         assertThat(result1).contains(member);
 
-
-        List<Member> result2 = memberRepository.findByUsername_Querydsl("memberdummy001");
-        System.out.println("result 2 : "+result2.toString());
-
-        assertThat(result2).containsExactly(member);
-    }
-
-    @Test
-    public void searchTest() {
-        Team teamA = new Team("teamX");
-        Team teamB = new Team("teamY");
-        em.persist(teamA);
-        em.persist(teamB);
-        Member member1 = new Member("d_member1", null, teamA,10);
-        Member member2 = new Member("d_member2", null, teamA,20);
-        Member member3 = new Member("d_member3", null, teamB,30);
-        Member member4 = new Member("d_member4", null, teamB,40);
-        em.persist(member1);
-        em.persist(member2);
-        em.persist(member3);
-        em.persist(member4);
         MemberSearchCondition condition = new MemberSearchCondition();
-        condition.setAgeGoe(35);
-        condition.setAgeLoe(40);
+        condition.setAgeGoe(40);
+        condition.setAgeLoe(50);
         condition.setTeamName("teamY");
-        List<MemberTeamDto> result =
-                memberRepository.searchByBuilder(condition);
-        System.out.println("result:"+result.toString());
-//        assertThat(result).extracting("username").containsExactly("d_member4");
 
-        List<MemberTeamDto> result2 =
-                memberRepository.search(condition);
-        System.out.println("result:"+result2.toString());
-//        assertThat(result2).extracting("username").containsExactly("d_member4");
+        List<MemberTeamDto> result2 = memberRepository.search(condition);
+        System.out.println("result2:"+result2.toString());
+
+        assertThat(result2).extracting("username").contains("d_member_47");
 
     }
 }
