+++ src/test/java/jpabook/jpashop_v2/service/MemberServiceTest.java
@@ -1,11 +1,10 @@
 package jpabook.jpashop_v2.service;
 
 import jpabook.jpashop_v2.domain.Member;
-import jpabook.jpashop_v2.repository.MemberRepository;
+import jpabook.jpashop_v2.repository.MemberJpaRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.Rollback;
 import org.springframework.transaction.annotation.Transactional;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -16,7 +15,8 @@ class MemberServiceTest
 {
 
     @Autowired MemberService memberService;
-    @Autowired MemberRepository memberRepository;
+    @Autowired
+    MemberJpaRepository memberRepository;
 
 
     @Test
